[#238] hotfix: 패키지 변경 및 DebateRepository.findByIdWithContributeWithoutAmendment 생성
