+++ src/test/java/goorm/eagle7/stelligence/domain/notification/listener/ContributeRejectedNotifierTest.java
@@ -21,6 +21,7 @@ import goorm.eagle7.stelligence.domain.document.content.model.Document;
 import goorm.eagle7.stelligence.domain.member.model.Member;
 import goorm.eagle7.stelligence.domain.notification.NotificationRequest;
 import goorm.eagle7.stelligence.domain.notification.NotificationSender;
+import goorm.eagle7.stelligence.domain.notification.listener.contribute.ContributeRejectedNotifier;
 import goorm.eagle7.stelligence.domain.vote.VoteRepository;
 
 @ExtendWith(MockitoExtension.class)
